+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/fargate/utils/estimate_utils.py
@@ -1,13 +1,8 @@
-from typing import Final
-
 import polars as pl
 
 from polars_utils import utils
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
 from utils.column_values.categorical_column_values import MainJobRoleLabels
-
-# Define constants for IDs for original length data.
-ROW_ID: Final[str] = "id"
 
 
 def calculate_estimated_filled_posts_by_job_role(lf: pl.LazyFrame) -> pl.LazyFrame:
@@ -109,7 +104,7 @@ def calculate_reg_man_difference(lf: pl.LazyFrame) -> pl.LazyFrame:
     posts and count of registered managers from CQC.
 
     The difference is calculated on registered manager rows, and the result is
-    copied to all rows, per locationid/import_date ("id").
+    copied to all rows, per locationid and cqc_location_import_date.
 
     Args:
         lf (pl.LazyFrame): A LazyFrame with columns
@@ -131,7 +126,7 @@ def calculate_reg_man_difference(lf: pl.LazyFrame) -> pl.LazyFrame:
                 == MainJobRoleLabels.registered_manager
             )
             .first(ignore_nulls=True)
-            .over(ROW_ID)
+            .over(IndCQC.id_per_locationid_import_date)
         ).alias(IndCQC.difference_between_estimate_and_cqc_registered_managers)
     )
 
@@ -158,11 +153,14 @@ def calculate_non_rm_managerial_distribution(
         pl.col(IndCQC.estimate_filled_posts_by_job_role)
         .filter(non_rm_manager_condition)
         .sum()
-        .over(ROW_ID)
+        .over(IndCQC.id_per_locationid_import_date)
     )
 
     count_non_rm_managerial_roles_expr = (
-        pl.lit(1).filter(non_rm_manager_condition).sum().over(ROW_ID)
+        pl.lit(1)
+        .filter(non_rm_manager_condition)
+        .sum()
+        .over(IndCQC.id_per_locationid_import_date)
     )
 
     lf = lf.with_columns(
@@ -253,9 +251,18 @@ def calc_diff_estimate_filled_posts_and_from_all_job_roles(
     posts_by_job_role_col = IndCQC.estimate_filled_posts_by_job_role_manager_adjusted
 
     sum_expr = (
-        pl.when(pl.col(posts_by_job_role_col).count().over(ROW_ID) == 0)
+        pl.when(
+            pl.col(posts_by_job_role_col)
+            .count()
+            .over(IndCQC.id_per_locationid_import_date)
+            == 0
+        )
         .then(None)
-        .otherwise(pl.col(posts_by_job_role_col).sum().over(ROW_ID))
+        .otherwise(
+            pl.col(posts_by_job_role_col)
+            .sum()
+            .over(IndCQC.id_per_locationid_import_date)
+        )
     )
 
     return lf.with_columns(